Moved Counting Helpers to the new countHelper.js file
